fix(evals): address code review issues in eval config dialogs
- Extract shared useAllConnectorTools, ConnectorToolSelect, AssertionRowsEditor,
  AssertionRow, OP_OPTIONS, and newRowId into eval-config-form-parts.tsx, removing
  ~170 lines of duplication between the edit and picker dialogs
- Replace newRowId() module-level counter with crypto.randomUUID() for safer key gen
- Replace AnyTestCaseOverride index signature + 'as' casts in eval-suites.routes.ts
  with a proper TestCaseOverrideRow type with optional join fields
- Add TODO comments on the shadow 'name' column in both junction tables
- Document non-atomic clone flow in evaluators-panel and test-cases-panel
- Add comment explaining why evaluator DELETE/PATCH use eval_suites:create permission
